catalogo marcas, listo

+++ be_/prcd/eliminar_catalogo.php
@@ -14,9 +14,11 @@
 <?php
 include('../../query/qconn/qc.php');
 
-$id = $_REQUEST['id'];
 
+$tipo = $_REQUEST['type'];
 
+if($tipo == 1){
+    $id = $_REQUEST['id'];
     $sqlDelete = "DELETE FROM color WHERE id='$id'";
     $resultado = $conn->query($sqlDelete);
 
@@ -37,6 +39,31 @@ $id = $_REQUEST['id'];
         else{
         echo 'No se registró producto';
         }
+}
+
+if($tipo == 2){
+    $id = $_REQUEST['id'];
+    $sqlDelete = "DELETE FROM marcas WHERE id='$id'";
+    $resultado = $conn->query($sqlDelete);
+
+    if($resultado){
+        
+        echo "<script type=\"text/javascript\">
+        Swal.fire({
+            icon: 'success',
+            imageUrl: '../../assets/brand/img/logo_store_shoes_sin_fondo.png',
+            imageHeight: 200,
+            imageAlt: 'Shoes Store Mx',
+            title: 'Marca eliminada',
+            text: 'Catálogo actualizado',
+            confirmButtonColor: '#3085d6',
+            footer: 'Shoes Store Mx'
+        }).then(function(){window.location='../categorias.php';});</script>";
+        }
+        else{
+        echo 'No se registró producto';
+        }
+}
 ?>
 
 </body>
